--- a/src/ACEmd.jl
+++ b/src/ACEmd.jl
@@ -10,7 +10,7 @@ using Folds
 @reexport using Folds: ThreadedEx, SequentialEx, DistributedEx
 using FLoops
 using NeighbourLists
-#using SparseArrays
+using SparseArrays
 using StaticArrays
 @reexport using Unitful
 @reexport using UnitfulAtomic