--- a/src/api.jl
+++ b/src/api.jl
@@ -1,15 +1,15 @@
 """
-    ace_energy(potential, ACE1.Atoms, Kwargs)
+    ace_energy(potential, atoms, Kwargs)
 
 Calculates ACE potential energy for atomic system.
-The `ACE1.Atoms` object needs to be in `ACE1.AtomsBase` compatable format.
+The `atoms` object needs to be in `AtomsBase` compatable format.
 The returned energy has a unit as defined by `Unitful`.
 
 Parallel execution is done with Transducers.jl and there is an option to
 use different executors. Look for `ThreadedEx` for more details on how to control it. 
 
 # Kwargs
-- `domain=1:length(ACE1.Atoms)`  :  choose subset of ACE1.Atoms to which energy is calculated
+- `domain=1:length(atoms)`  :  choose subset of atoms to which energy is calculated
 - `executor=ThreadedEx()`   :  used to control multithreading using Transducers.jl
 - `energy_unit`  :   used to override energy unit for the calculation
 - `length_unit`  :   used to override lenght unit for the calculation
@@ -96,17 +96,17 @@ end
 ## forces
 
 """
-    ace_forces(potential, ACE1.Atoms, Kwargs)
+    ace_forces(potential, atoms, Kwargs)
 
 Calculates forces for ACE potential for given atomic system.
-The `ACE1.Atoms` object needs to be in `ACE1.AtomsBase` compatable format.
+The `atoms` object needs to be in `AtomsBase` compatable format.
 The returned energy has a unit as defined by `Unitful`.
 
 Parallel execution is done with Transducers.jl and there is an option to
 use different executors. Look for `ThreadedEx` for more details on how to control it. 
 
 # Kwargs
-- `domain=1:length(ACE1.Atoms)`  :  choose subset of ACE1.Atoms to which energy is calculated
+- `domain=1:length(atoms)`  :  choose subset of atoms to which energy is calculated
 - `executor=ThreadedEx()`   :  used to control multithreading using Transducers.jl
 - `energy_unit`  :   used to override energy unit for the calculation
 - `length_unit`  :   used to override lenght unit for the calculation
@@ -121,17 +121,14 @@ function ace_forces(V, at;
         kwargs...
     )
     nlist = neighborlist(at, get_cutoff(V; cutoff_unit=cutoff_unit) )
-    F = Folds.sum( domain, executor ) do i
+    F = Folds.sum( domain, executor; init=zeros(SVector{3, Float64}, length(at)) ) do i
         j, R, Z = neigsz(nlist, at, i)
         _, tmp = ace_evaluate_d(V, R, Z, _atomic_number(at,i))
 
-        #TODO make this faster
-        f = zeros(eltype(tmp.dV), length(at))
-        for k in eachindex(j)
-            f[j[k]] -= tmp.dV[k]
-            f[i]    += tmp.dV[k]
-        end
-        f
+        # Make a copy with -1 .* tmp.dV
+        sout = sparsevec( [j[k] for k in eachindex(j)], -1 .* tmp.dV, length(at) )
+        sout[i] += sum(tmp.dV)
+        sout
     end
     return F * (energy_unit / length_unit)
 end
@@ -153,17 +150,17 @@ end
 ## virial
 
 """
-    ace_virial(potential, ACE1.Atoms, Kwargs)
+    ace_virial(potential, atoms, Kwargs)
 
 Calculates virial for ACE potential for given atomic system.
-The `ACE1.Atoms` object needs to be in `ACE1.AtomsBase` compatable format.
+The `atoms` object needs to be in `AtomsBase` compatable format.
 The returned energy has a unit as defined by `Unitful`.
 
 Parallel execution is done with Transducers.jl and there is an option to
 use different executors. Look for `ThreadedEx` for more details on how to control it. 
 
 # Kwargs
-- `domain=1:length(ACE1.Atoms)`  :  choose subset of ACE1.Atoms to which energy is calculated
+- `domain=1:length(atoms)`  :  choose subset of atoms to which energy is calculated
 - `executor=ThreadedEx()`   :  used to control multithreading using Transducers.jl
 - `energy_unit`  :   used to override energy unit for the calculation
 - `length_unit`  :   used to override lenght unit for the calculation