This adds small performance improvement. By changing a temporary array to SparseArray type, which speeds up force calculations for large system.