@tjjarvinen 2023-08-26
Reading benchmark results:

This is new
```
                                ID            time    GC time          memory allocations
  ––––––––––––––––––––––––––––––– ––––––––––––––– –––––––––– ––––––––––––––– –––––––––––
   ["AtomsBase", "energy", "big"]  99.881 ms (5%)             74.44 MiB (1%)     1255580
  ["AtomsBase", "energy", "huge"]    1.866 s (5%) 323.196 ms   1.08 GiB (1%)    19629264
   ["AtomsBase", "forces", "big"] 228.836 ms (5%)  16.448 ms 215.45 MiB (1%)     2786528
  ["AtomsBase", "forces", "huge"]    5.943 s (5%) 961.246 ms  11.16 GiB (1%)    43597538
   ["AtomsBase", "virial", "big"] 176.475 ms (5%)            130.94 MiB (1%)     2722605
  ["AtomsBase", "virial", "huge"]    3.342 s (5%) 562.743 ms   1.94 GiB (1%)    42544027
       ["JuLIP", "energy", "big"]  53.254 ms (5%)             52.83 MiB (1%)     1191925
      ["JuLIP", "energy", "huge"]    3.535 s (5%) 283.899 ms   2.59 GiB (1%)    18638120
       ["JuLIP", "forces", "big"] 113.772 ms (5%)             88.13 MiB (1%)     2629828
      ["JuLIP", "forces", "huge"]    2.300 s (5%) 391.639 ms   1.34 GiB (1%)    41088431
       ["JuLIP", "virial", "big"] 105.031 ms (5%)             87.87 MiB (1%)     2629809
      ["JuLIP", "virial", "huge"]    2.286 s (5%) 415.066 ms   1.34 GiB (1%)    41088413
```

and this is old

```
                                ID            time    GC time          memory allocations
  ––––––––––––––––––––––––––––––– ––––––––––––––– –––––––––– ––––––––––––––– –––––––––––
   ["AtomsBase", "energy", "big"] 102.329 ms (5%)             74.44 MiB (1%)     1255579
  ["AtomsBase", "energy", "huge"]    1.910 s (5%) 306.197 ms   1.08 GiB (1%)    19629264
   ["AtomsBase", "forces", "big"] 242.877 ms (5%)  34.504 ms 237.41 MiB (1%)     3123813
  ["AtomsBase", "forces", "huge"]    7.644 s (5%)    1.385 s  25.01 GiB (1%)    50261479
   ["AtomsBase", "virial", "big"] 180.286 ms (5%)            130.94 MiB (1%)     2722604
  ["AtomsBase", "virial", "huge"]    3.187 s (5%) 507.721 ms   1.94 GiB (1%)    42544026
       ["JuLIP", "energy", "big"]  51.041 ms (5%)             52.83 MiB (1%)     1191926
      ["JuLIP", "energy", "huge"]    3.561 s (5%) 294.893 ms   2.59 GiB (1%)    18638120
       ["JuLIP", "forces", "big"]    9.143 s (5%) 381.064 ms 642.18 MiB (1%)    10830509
      ["JuLIP", "forces", "huge"]    2.540 s (5%) 567.630 ms   1.34 GiB (1%)    41088431
       ["JuLIP", "virial", "big"] 108.140 ms (5%)             87.87 MiB (1%)     2629810
      ["JuLIP", "virial", "huge"]    2.242 s (5%) 407.436 ms   1.34 GiB (1%)    41088413
```

So in summary, performance is defined by memory allocations and where even those can be reduced it results in increased performance.