@tjjarvinen 2023-08-28
Benchmark summary

New

```
                               ID            time    GC time          memory allocations
  ––––––––––––––––––––––––––––––– ––––––––––––––– –––––––––– ––––––––––––––– –––––––––––
   ["AtomsBase", "energy", "big"]  88.689 ms (5%)             74.44 MiB (1%)     1255579
  ["AtomsBase", "energy", "huge"]    1.768 s (5%) 302.026 ms   1.08 GiB (1%)    19629264
   ["AtomsBase", "forces", "big"] 198.228 ms (5%)  28.931 ms 135.53 MiB (1%)     2893260
  ["AtomsBase", "forces", "huge"]    3.399 s (5%) 651.583 ms   2.02 GiB (1%)    45572291
   ["AtomsBase", "virial", "big"] 170.749 ms (5%)            130.94 MiB (1%)     2722605
  ["AtomsBase", "virial", "huge"]    3.172 s (5%) 600.776 ms   1.94 GiB (1%)    42544027
       ["JuLIP", "energy", "big"]  50.520 ms (5%)             52.83 MiB (1%)     1191925
      ["JuLIP", "energy", "huge"]    3.612 s (5%) 302.136 ms   2.59 GiB (1%)    18638120
       ["JuLIP", "forces", "big"] 102.100 ms (5%)             88.13 MiB (1%)     2629827
      ["JuLIP", "forces", "huge"]    2.190 s (5%) 410.440 ms   1.34 GiB (1%)    41088431
       ["JuLIP", "virial", "big"]  99.146 ms (5%)             87.87 MiB (1%)     2629809
      ["JuLIP", "virial", "huge"]    2.207 s (5%) 413.153 ms   1.34 GiB (1%)    41088413
```

Old

```
                                ID            time    GC time          memory allocations
  ––––––––––––––––––––––––––––––– ––––––––––––––– –––––––––– ––––––––––––––– –––––––––––
   ["AtomsBase", "energy", "big"] 105.593 ms (5%)             74.44 MiB (1%)     1255579
  ["AtomsBase", "energy", "huge"]    1.819 s (5%) 337.992 ms   1.08 GiB (1%)    19629264
   ["AtomsBase", "forces", "big"] 244.741 ms (5%)  44.578 ms 237.41 MiB (1%)     3123813
  ["AtomsBase", "forces", "huge"]    7.187 s (5%)    1.311 s  25.01 GiB (1%)    50261479
   ["AtomsBase", "virial", "big"] 168.682 ms (5%)            130.94 MiB (1%)     2722604
  ["AtomsBase", "virial", "huge"]    3.277 s (5%) 630.443 ms   1.94 GiB (1%)    42544026
       ["JuLIP", "energy", "big"]  47.667 ms (5%)             52.83 MiB (1%)     1191926
      ["JuLIP", "energy", "huge"]    3.644 s (5%) 286.099 ms   2.59 GiB (1%)    18638119
       ["JuLIP", "forces", "big"]    9.514 s (5%) 357.902 ms 642.35 MiB (1%)    10836532
      ["JuLIP", "forces", "huge"]    2.139 s (5%) 404.638 ms   1.34 GiB (1%)    41088431
       ["JuLIP", "virial", "big"] 103.352 ms (5%)             87.87 MiB (1%)     2629810
      ["JuLIP", "virial", "huge"]    2.153 s (5%) 370.208 ms   1.34 GiB (1%)    41088413
```

There seems to be some odd results, like old code for forces on "big" system, so take them with grain of salt.